UI 연동완료 했는데 몬스터 죽었을때 EXP올라가는거, EXP올라가서 4개차면 레벨업하게 해야됌

+++ Team_Mockup_Project-Dungreed/UiAbility.h
@@ -51,10 +51,10 @@ public:
 	void SetCriticalDamage(int criticalDamage);
 	void SetCriticalPercent(int criticalPercent);
 	void SetAttackSpeed(float attackSpeed);
-	void SetMoveMentSpeed(float moveSpeed);
+	void SetMoveMentSpeed(int moveSpeed);
 	void SetArmor(int armor);
 	void SetArmorPercent(int armorPercent);
-	void SetDashDamagePercent(float dashDamage);
+	void SetDashDamagePercent(int dashDamage);
 };
 /*
 * float attackDamage;
